+++ src/core_services/apple_events/mod.rs
@@ -1,7 +1,9 @@
 mod desc;
 mod event_class;
 mod event_id;
+mod return_id;
 
 pub use desc::*;
 pub use event_class::*;
 pub use event_id::*;
+pub use return_id::*;