+++ CHANGELOG.md
@@ -116,6 +116,8 @@ The format is based on [Keep a Changelog] and this project adheres to
     - `AEDescType`, `AEEventClass`, and `AEEventID` newtype structs over
       `FourCharCode`.
 
+    - `AEReturnID` newtype struct over `i16`.
+
 - Created `core_animation` module for
   [Core Animation](https://developer.apple.com/documentation/quartzcore)
   framework.